Split the app view panel switching and lock handling into mixins.
